requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
requests>=2.31
python-dotenv>=1.0
tqdm>=4.66
pydantic>=2.5
joblib>=1.3
rich>=13.7
anthropic>=0.80.0
nba_api>=1.4